requirements.txt
Flask[async]==3.0.3
requests
google-cloud-firestore
python-dotenv
simple-salesforce==1.12.6
google-cloud-logging
pytest 
pytest-mock
datetime
aiohttp
httpx